Add some metrics for python udf plans.
Along with some refactor to reduce the duplicate code.
